Move functions to builder
